backend/requirements.txt
fastapi
python-socketio[asyncio]
sqlalchemy
